+++ src/ska_mid_cbf_tdc_mcs/fsp/fsp_corr_subarray_component_manager.py
@@ -15,19 +15,18 @@ from threading import Event
 from typing import Any, Callable, Optional
 
 import tango
-from ska_control_model import CommunicationStatus, PowerState, TaskStatus
+from ska_control_model import CommunicationStatus, TaskStatus
 from ska_tango_base.commands import ResultCode
-from ska_tango_testing import context
 
 from ska_mid_cbf_tdc_mcs.commons.global_enum import const, freq_band_dict
-from ska_mid_cbf_tdc_mcs.component.obs_component_manager import (
-    CbfObsComponentManager,
+from ska_mid_cbf_tdc_mcs.fsp.fsp_mode_subarray_component_manager import (
+    FspModeSubarrayComponentManager,
 )
 
 FSP_CORR_PARAM_PATH = "mnt/fsp_param/internal_params_fsp_corr_subarray.json"
 
 
-class FspCorrSubarrayComponentManager(CbfObsComponentManager):
+class FspCorrSubarrayComponentManager(FspModeSubarrayComponentManager):
     """
     A component manager for the FspCorrSubarray device.
     """
@@ -35,26 +34,20 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
     def __init__(
         self: FspCorrSubarrayComponentManager,
         *args: Any,
-        hps_fsp_corr_controller_fqdn: str,
         **kwargs: Any,
     ) -> None:
         """
         Initialise a new instance.
 
-        # TODO: for Mid.CBF, param hps_fsp_corr_controller_fqdn to be updated to a list of FQDNs (max length = 20), one entry for each Talon board in the FSP_UNIT
         :param hps_fsp_corr_controller_fqdn: FQDN of the HPS FSP Correlator controller device
         """
-        super().__init__(*args, **kwargs)
+        super().__init__(
+            *args,
+            **kwargs,
+        )
 
-        self._hps_fsp_corr_controller_fqdn = hps_fsp_corr_controller_fqdn
-        self._proxy_hps_fsp_corr_controller = None
-
-        self.delay_model = ""
-        self.vcc_ids = []
         self.frequency_band = 0
         self.frequency_slice_id = 0
-        self.scan_id = 0
-        self.config_id = ""
         self.channel_averaging_map = [
             [
                 int(
@@ -75,81 +68,9 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
 
         self.output_link_map = [[0, 0] for _ in range(40)]
 
-        self.last_hps_scan_configuration = ""
-
-    # -------------
-    # Communication
-    # -------------
-
-    def _start_communicating(
-        self: FspCorrSubarrayComponentManager, *args, **kwargs
-    ) -> None:
-        """
-        Establish communication with the component, then start monitoring.
-        """
-        # Try to connect to HPS devices, which are deployed during the
-        # CbfController OnCommand sequence
-        if not self.simulation_mode:
-            try:
-                self._proxy_hps_fsp_corr_controller = context.DeviceProxy(
-                    device_name=self._hps_fsp_corr_controller_fqdn
-                )
-                self._proxy_hps_fsp_corr_controller.set_timeout_millis(
-                    self._lrc_timeout * 1000
-                )
-            except tango.DevFailed as df:
-                self.logger.error(
-                    f"Failed to connect to {self._hps_fsp_corr_controller_fqdn}; {df}"
-                )
-                self._update_communication_state(
-                    communication_state=CommunicationStatus.NOT_ESTABLISHED
-                )
-                return (
-                    ResultCode.FAILED,
-                    "Failed to establish proxy to HPS FSP Corr controller device.",
-                )
-
-        super()._start_communicating()
-        self._update_component_state(power=PowerState.ON)
-
     # -------------
     # Class Helpers
     # -------------
-
-    def _assign_vcc(
-        self: FspCorrSubarrayComponentManager, argin: list[int]
-    ) -> None:
-        """
-        Assign specified VCCs to the FSP CORR subarray.
-
-        :param argin: IDs of VCCs to add.
-        """
-
-        for vccID in argin:
-            if vccID not in self.vcc_ids:
-                self.logger.info(f"VCC {vccID} assigned.")
-                self.vcc_ids.append(vccID)
-            else:
-                log_msg = f"VCC {vccID} already assigned to current FSP CORR subarray."
-                self.logger.warning(log_msg)
-
-    def _release_vcc(
-        self: FspCorrSubarrayComponentManager, argin: list[int]
-    ) -> None:
-        """
-        Release assigned VCC from the FSP CORR subarray.
-
-        :param argin: IDs of VCCs to remove.
-        """
-        for vccID in argin:
-            if vccID in self.vcc_ids:
-                self.logger.info(f"VCC {vccID} released.")
-                self.vcc_ids.remove(vccID)
-            else:
-                log_msg = (
-                    "VCC {vccID} not assigned to FSP CORR subarray. Skipping."
-                )
-                self.logger.warning(log_msg)
 
     def _build_hps_fsp_config(
         self: FspCorrSubarrayComponentManager, configuration: dict
@@ -196,50 +117,12 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
         """Deconfigure scan configuration parameters."""
         self.frequency_band = 0
         self.frequency_slice_id = 0
-        self.scan_id = 0
-        self.config_id = ""
-        self.last_hps_scan_configuration = ""
 
-        # release all assigned VCC to reset to IDLE state
-        self._release_vcc(self.vcc_ids.copy())
+        super()._deconfigure()
 
     # -------------
     # Fast Commands
     # -------------
-
-    def update_delay_model(
-        self: FspCorrSubarrayComponentManager, model: str
-    ) -> tuple[ResultCode, str]:
-        """
-        Update the FSP's delay model (serialized JSON object)
-
-        :param model: the delay model data
-        :return: A tuple containing a return code and a string
-                message indicating status. The message is for
-                information purpose only.
-        :rtype: (ResultCode, str)
-        """
-        self.logger.info("Entering FspCorrSubarray.update_delay_model()")
-
-        if not self.simulation_mode:
-            try:
-                self._proxy_hps_fsp_corr_controller.UpdateDelayModels(model)
-            except tango.DevFailed as df:
-                self.logger.error(f"{df}")
-                self._update_communication_state(
-                    communication_state=CommunicationStatus.NOT_ESTABLISHED
-                )
-                return (
-                    ResultCode.FAILED,
-                    "Failed to issue UpdateDelayModels command to HPS FSP Corr controller",
-                )
-
-        # the whole delay model must be stored
-        self.delay_model = model
-        self.device_attr_change_callback("delayModel", model)
-        self.device_attr_archive_callback("delayModel", model)
-
-        return (ResultCode.OK, "UpdateDelayModel completed OK")
 
     # ---------------------
     # Long Running Commands
@@ -255,6 +138,8 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
         Execute configure scan operation.
 
         :param argin: JSON string with the configure scan parameters
+        :param task_callback: command tracker update_command_info callback
+        :param task_abort_event: task executor abort event
 
         :return: None
         """
@@ -280,12 +165,11 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
         self._assign_vcc(configuration["corr_vcc_ids"])
 
         # Issue ConfigureScan to HPS FSP Corr controller
-
         if not self.simulation_mode:
             hps_fsp_configuration = self._build_hps_fsp_config(configuration)
             self.last_hps_scan_configuration = hps_fsp_configuration
             try:
-                self._proxy_hps_fsp_corr_controller.ConfigureScan(
+                self._proxy_hps_fsp_mode_controller.ConfigureScan(
                     hps_fsp_configuration
                 )
             except tango.DevFailed as df:
@@ -309,235 +193,6 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
 
         task_callback(
             result=(ResultCode.OK, "ConfigureScan completed OK"),
-            status=TaskStatus.COMPLETED,
-        )
-        return
-
-    def _scan(
-        self: FspCorrSubarrayComponentManager,
-        argin: int,
-        task_callback: Optional[Callable] = None,
-        task_abort_event: Optional[Event] = None,
-    ) -> None:
-        """
-        Begin scan operation.
-
-        :param argin: scan ID integer
-
-        :return: None
-        """
-        # set task status in progress, check for abort event
-        task_callback(status=TaskStatus.IN_PROGRESS)
-        if self.task_abort_event_is_set(
-            "Scan", task_callback, task_abort_event
-        ):
-            return
-
-        self.scan_id = argin
-
-        if not self.simulation_mode:
-            try:
-                self._proxy_hps_fsp_corr_controller.Scan(self.scan_id)
-            except tango.DevFailed as df:
-                self.logger.error(f"{df}")
-                self._update_communication_state(
-                    communication_state=CommunicationStatus.NOT_ESTABLISHED
-                )
-                task_callback(
-                    status=TaskStatus.FAILED,
-                    result=(
-                        ResultCode.FAILED,
-                        "Failed to issue Scan command to HPS FSP Corr controller device.",
-                    ),
-                )
-                return
-
-        # Update obsState callback
-        self._update_component_state(scanning=True)
-
-        task_callback(
-            result=(ResultCode.OK, "Scan completed OK"),
-            status=TaskStatus.COMPLETED,
-        )
-        return
-
-    def _end_scan(
-        self: FspCorrSubarrayComponentManager,
-        task_callback: Optional[Callable] = None,
-        task_abort_event: Optional[Event] = None,
-    ) -> None:
-        """
-        End scan operation.
-
-        :return: None
-        """
-        # set task status in progress, check for abort event
-        task_callback(status=TaskStatus.IN_PROGRESS)
-        if self.task_abort_event_is_set(
-            "EndScan", task_callback, task_abort_event
-        ):
-            return
-
-        if not self.simulation_mode:
-            try:
-                self._proxy_hps_fsp_corr_controller.EndScan()
-            except tango.DevFailed as df:
-                self.logger.error(f"{df}")
-                self._update_communication_state(
-                    communication_state=CommunicationStatus.NOT_ESTABLISHED
-                )
-                task_callback(
-                    status=TaskStatus.FAILED,
-                    result=(
-                        ResultCode.FAILED,
-                        "Failed to issue EndScan command to HPS FSP Corr controller device.",
-                    ),
-                )
-                return
-
-        # Update obsState callback
-        self._update_component_state(scanning=False)
-
-        task_callback(
-            result=(ResultCode.OK, "EndScan completed OK"),
-            status=TaskStatus.COMPLETED,
-        )
-        return
-
-    def _go_to_idle(
-        self: FspCorrSubarrayComponentManager,
-        task_callback: Optional[Callable] = None,
-        task_abort_event: Optional[Event] = None,
-    ) -> None:
-        """
-        Execute observing state transition from READY to IDLE.
-
-        :return: None
-        """
-        # set task status in progress, check for abort event
-        task_callback(status=TaskStatus.IN_PROGRESS)
-        if self.task_abort_event_is_set(
-            "GoToIdle", task_callback, task_abort_event
-        ):
-            return
-
-        if not self.simulation_mode:
-            try:
-                self._proxy_hps_fsp_corr_controller.GoToIdle()
-            except tango.DevFailed as df:
-                self.logger.error(f"{df}")
-                self._update_communication_state(
-                    communication_state=CommunicationStatus.NOT_ESTABLISHED
-                )
-                task_callback(
-                    status=TaskStatus.FAILED,
-                    result=(
-                        ResultCode.FAILED,
-                        "Failed to issue Unconfigure command to HPS FSP Corr controller device.",
-                    ),
-                )
-                return
-
-        # reset configured attributes
-        self._deconfigure()
-
-        # Update obsState callback
-        self._update_component_state(configured=False)
-
-        task_callback(
-            result=(ResultCode.OK, "GoToIdle completed OK"),
-            status=TaskStatus.COMPLETED,
-        )
-        return
-
-    def _abort(
-        self: FspCorrSubarrayComponentManager,
-        task_callback: Optional[Callable] = None,
-        task_abort_event: Optional[Event] = None,
-    ) -> None:
-        """
-        Abort the current scan operation.
-
-        :return: None
-        """
-        # set task status in progress, check for abort event
-        task_callback(status=TaskStatus.IN_PROGRESS)
-        if self.task_abort_event_is_set(
-            "Abort", task_callback, task_abort_event
-        ):
-            return
-        try:
-            # TODO: Abort command not implemented for the HPS FSP application
-            pass
-        except tango.DevFailed as df:
-            self.logger.error(f"{df}")
-            self._update_communication_state(
-                communication_state=CommunicationStatus.NOT_ESTABLISHED
-            )
-            task_callback(
-                status=TaskStatus.FAILED,
-                result=(
-                    ResultCode.FAILED,
-                    "Failed to issue Abort command to HPS FSP Corr controller device.",
-                ),
-            )
-            return
-
-        # Update obsState callback
-        self._update_component_state(scanning=False)
-
-        task_callback(
-            result=(ResultCode.OK, "Abort completed OK"),
-            status=TaskStatus.COMPLETED,
-        )
-        return
-
-    def _obs_reset(
-        self: FspCorrSubarrayComponentManager,
-        task_callback: Optional[Callable] = None,
-        task_abort_event: Optional[Event] = None,
-    ) -> None:
-        """
-        Reset the scan operation from ABORTED or FAULT.
-
-        :return: None
-        """
-        # set task status in progress, check for abort event
-        task_callback(status=TaskStatus.IN_PROGRESS)
-        if self.task_abort_event_is_set(
-            "ObsReset", task_callback, task_abort_event
-        ):
-            return
-
-        if not self.simulation_mode:
-            try:
-                pass
-                # TODO: ObsReset command not implemented for the HPS FSP application, see CIP-1850
-                # self._proxy_hps_fsp_corr_controller.ObsReset()
-            except tango.DevFailed as df:
-                self.logger.error(f"{df}")
-                self._update_communication_state(
-                    communication_state=CommunicationStatus.NOT_ESTABLISHED
-                )
-                task_callback(
-                    status=TaskStatus.FAILED,
-                    result=(
-                        ResultCode.FAILED,
-                        "Failed to issue ObsReset to HPS FSP Corr controller device.",
-                    ),
-                )
-                return
-
-        # reset configured attributes
-        self._deconfigure()
-
-        # Update obsState callback
-        # There is no obsfault == False action implemented, however,
-        # we reset it it False so that obsfault == True may be triggered in the future
-        self._update_component_state(configured=False, obsfault=False)
-
-        task_callback(
-            result=(ResultCode.OK, "ObsReset completed OK"),
             status=TaskStatus.COMPLETED,
         )
         return