+++ src/ska_mid_cbf_tdc_mcs/fsp/fsp_multi.py
@@ -13,6 +13,7 @@ from tango.server import run
 
 from ska_mid_cbf_tdc_mcs.fsp.fsp_corr_subarray_device import FspCorrSubarray
 from ska_mid_cbf_tdc_mcs.fsp.fsp_device import Fsp
+from ska_mid_cbf_tdc_mcs.fsp.fsp_pst_subarray_device import FspPstSubarray
 
 file_path = os.path.dirname(os.path.abspath(__file__))
 
@@ -21,7 +22,15 @@ __all__ = ["main"]
 
 
 def main(args=None, **kwargs):
-    return run(classes=(FspCorrSubarray, Fsp), args=args, **kwargs)
+    return run(
+        classes=(
+            FspCorrSubarray,
+            FspPstSubarray,
+            Fsp,
+        ),
+        args=args,
+        **kwargs
+    )
 
 
 if __name__ == "__main__":