+++ src/ska_mid_cbf_tdc_mcs/visibility_transport/visibility_transport.py
@@ -5,6 +5,7 @@ routing the visibilities from FSPs to SDP.
 It is assumed that TalonDX boards will only be used in Mid-CBF up to AA1,
 supporting up to 8 boards.
 """
+
 import ctypes
 import logging