+++ tests/unit/fsp/Fsp_test.py
@@ -63,6 +63,7 @@ class TestFsp:
             ),
             HpsFspControllerAddress="talondx-001/fsp-app/fsp-controller",
             HpsFspCorrControllerAddress="talondx-001/fsp-app/fsp-corr-controller",
+            HpsFspPstControllerAddress="talondx-001/fsp-app/fsp-pst-controller",
             DeviceID="1",
         )
         for name, mock in initial_mocks.items():
@@ -216,7 +217,7 @@ class TestFsp:
             ),
         )
 
-    @pytest.mark.parametrize("function_mode", [FspModes.CORR])
+    @pytest.mark.parametrize("function_mode", [FspModes.CORR, FspModes.PST_BF])
     def test_SetFunctionMode_not_allowed_from_off(
         self: TestFsp,
         device_under_test: context.DeviceProxy,
@@ -253,7 +254,7 @@ class TestFsp:
             ),
         )
 
-    @pytest.mark.parametrize("function_mode", [FspModes.CORR])
+    @pytest.mark.parametrize("function_mode", [FspModes.CORR, FspModes.PST_BF])
     def test_SetFunctionMode_not_allowed_already_set(
         self: TestFsp,
         device_under_test: context.DeviceProxy,
@@ -273,6 +274,7 @@ class TestFsp:
             device_under_test=device_under_test,
             event_tracer=event_tracer,
             sub_ids=[1, 2, 3],
+            fsp_mode=function_mode,
         )
 
         # test issuing SetFunctionMode on a previously set FSP
@@ -294,6 +296,10 @@ class TestFsp:
             ),
         )
 
+    @pytest.mark.parametrize(
+        "fsp_mode",
+        [FspModes.CORR, FspModes.PST_BF],
+    )
     # parameterized with all possible subarray IDs, a duplicate ID and IDs below and above range
     @pytest.mark.parametrize(
         "sub_ids",
@@ -308,6 +314,7 @@ class TestFsp:
         device_under_test: context.DeviceProxy,
         event_tracer: TangoEventTracer,
         sub_ids: list[int],
+        fsp_mode: FspModes,
     ) -> None:
         """
         Test the AddSubarrayMembership() command's happy path.
@@ -316,6 +323,7 @@ class TestFsp:
         :param event_tracer: A TangoEventTracer used to receive subscribed change
                              events from the device under test.
         :param sub_ids: list of subarray IDs to add
+        :param fsp_mode: FspMode to be set for the device
         """
         # set set function mode to CORR
         self.test_SetFunctionMode(
@@ -417,12 +425,17 @@ class TestFsp:
                 ),
             )
 
+    @pytest.mark.parametrize(
+        "fsp_mode",
+        [FspModes.CORR, FspModes.PST_BF],
+    )
     @pytest.mark.parametrize("sub_ids", [[1, 2, 3]])
     def test_RemoveSubarrayMembership(
         self: TestFsp,
         device_under_test: context.DeviceProxy,
         event_tracer: TangoEventTracer,
         sub_ids: list[int],
+        fsp_mode: FspModes,
     ) -> None:
         """
         Test the RemoveSubarrayMembership() command's happy path.
@@ -431,11 +444,12 @@ class TestFsp:
         :param event_tracer: A TangoEventTracer used to receive subscribed change
                              events from the device under test.
         :param sub_ids: list of subarray IDs to remove
+        :param fsp_mode: FspMode to be set for the device
         """
 
         # set device ONLINE, ON, function mode to CORR and add subarray membership
         self.test_AddSubarrayMembership(
-            device_under_test, event_tracer, sub_ids
+            device_under_test, event_tracer, sub_ids, fsp_mode
         )
 
         # test invalid subarray ID