+++ tests/unit/fsp/conftest.py
@@ -89,16 +89,35 @@ def mock_hps_fsp_corr_controller() -> unittest.mock.Mock:
 
 
 @pytest.fixture()
+def mock_hps_fsp_pst_controller() -> unittest.mock.Mock:
+    builder = MockDeviceBuilder()
+    builder.add_command("SetFunctionMode", None)
+    return builder()
+
+
+@pytest.fixture()
+def mock_hps_fsp_controller() -> unittest.mock.Mock:
+    builder = MockDeviceBuilder()
+    builder.add_command("SetFunctionMode", None)
+    return builder()
+
+
+@pytest.fixture()
 def initial_mocks(
     mock_fsp_corr_subarray_device: unittest.mock.Mock,
     mock_fsp_pst_subarray_device: unittest.mock.Mock,
     mock_hps_fsp_corr_controller: unittest.mock.Mock,
+    mock_hps_fsp_pst_controller: unittest.mock.Mock,
+    mock_hps_fsp_controller: unittest.mock.Mock,
 ) -> dict[str, unittest.mock.Mock]:
     """
     Return a dictionary of device proxy mocks to pre-register.
 
     :param mock_fsp_corr_subarray_device: a mock FspCorrSubarray.
     :param mock_fsp_pst_subarray_device: a mock FspPstSubarray.
+    :param mock_hps_fsp_corr_controller: a mock HPS FSP Corr Controller
+    :param mock_hps_fsp_pst_controller: a mock HPS FSP PST Controller
+    :param mock_hps_fspcontroller: a mock HPS FSP Top Level Controller
     :return: a dictionary of device proxy mocks to pre-register.
     """
     mocks = {}
@@ -109,5 +128,11 @@ def initial_mocks(
         mocks[
             f"mid_csp_cbf/fspPstSubarray/01_{sub_id:02}"
         ] = mock_fsp_pst_subarray_device
-    mocks["talondx-001/fsp-app/fsp-controller"] = mock_hps_fsp_corr_controller
+    mocks["talondx-001/fsp-app/fsp-controller"] = mock_hps_fsp_controller
+    mocks[
+        "talondx-001/fsp-app/fsp-corr-controller"
+    ] = mock_hps_fsp_corr_controller
+    mocks[
+        "talondx-001/fsp-app/fsp-pst-controller"
+    ] = mock_hps_fsp_pst_controller
     return mocks