+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ This project adheres to `Semantic Versioning http://semver.org/>`_.
 
 UNRELEASED CHANGES
 ******************
+* CIP-2660 Added back FspPstSubarray to MCS
+
+1.1.1
+*****
 * CIP-2953 increase SV pod storage from 375Mi to 2Gi
 * CIP-2659: CbfController On/Off command partial success
   * On command succeeds if one or more LRUs are powered on; Off command succeeds if one or more LRUs are powered off
@@ -24,6 +28,8 @@ UNRELEASED CHANGES
 * CIP-2809: replaced all custom assertions with ska-tango-testing assertions, deleted custom assertions
 * SKB-668: Sort VCC IDs by their associated receptor IDs when passing to FSPs
 * SKB-669: Added configurable timeout to values for TalonLRU's PowerSwitch proxies
+* CIP-3159: disable loopback on unused visibility SLIM rx devices to prevent unexpected visibility outputs from other FSPs
+* CIP-2790: Added error handling for CbfFastCommands processed within MCS
 
 1.1.0
 **********